// packages/race-management/src/webparts/raceMeetings/components/RaceMeetings.module.scss.d.ts
declare const styles: {
    raceMeetings: string;
    teams: string;
    calendarHeader: string;
    dateDisplay: string;
    filters: string;
    loading: string;
    calendarContent: string;
    dayView: string;
    timeGrid: string;
    hourSlot: string;
    hourLabel: string;
    hourContent: string;
    weekView: string;
    weekGrid: string;
    dayColumn: string;
    dayHeader: string;
    dayName: string;
    dayNumber: string;
    dayContent: string;
    monthView: string;
    weekDays: string;
    weekDay: string;
    week: string;
    day: string;
    otherMonth: string;
    today: string;
    meetings: string;
    meetingRectangle: string;
    trackName: string;
    moreMeetings: string;
    noMeetings: string;
    meetingsList: string;
    meetingCard: string;
    meetingHeader: string;
    meetingTitle: string;
    meetingAuthority: string;
    meetingDetails: string;
    meetingType: string;
    meeting: string;
    meetingTime: string;
    legend: string;
    legendItems: string;
    legendItem: string;
    legendColor: string;
    authorityBadge: string;
    dark: string;
};
export default styles;
//# sourceMappingURL=RaceMeetings.module.scss.d.ts.map