// packages/race-management/src/webparts/raceMeetings/components/IRaceMeetingsProps.d.ts
import { WebPartContext } from '@microsoft/sp-webpart-base';
import { CalendarView } from '../../../models/IRaceMeeting';
export interface IRaceMeetingsProps {
    isDarkTheme: boolean;
    environmentMessage: string;
    hasTeamsContext: boolean;
    userDisplayName: string;
    context: WebPartContext;
    defaultView: CalendarView;
    selectedAuthority: string;
    selectedTrackId: string;
    showPastMeetings: boolean;
    showFutureMeetings: boolean;
    onUpdateFilters?: (authority: string, trackId: string) => void;
}
//# sourceMappingURL=IRaceMeetingsProps.d.ts.map